backend/requirements.txt
numpy>=1.24.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2 